Add returnPressed signal to spin boxes
This differs from `editingFinished`, which is also triggered when the
focus is lost. `returnPressed` is useful if the user is to explicitly
confirm the acceptance of the value, for example by means of an Apply
button next to the SpinBox.

The corresponding signal of the internal QLineEdit cannot be used as
access to it is protected.

Change-Id: Ie9a146212e6aeac5cc9e4592b1633156323426dd
Fixes: QTBUG-99797
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>