ucstricmp8: fix Coverity ARRAY_VS_SINGLETON issue
Coverity has this checker where it complains if you use a T as a
T[1]. The C++ standard says this is fine¹, but qstring.cpp,
specifically, is security-critical, so we shouldn't leave Coverity
issues unfixed in there.

So replace uc1 with a decoded[1] array and make both output and uc1
point to decoded's first (and only) element.

Amends 45838673df6e64a6fd42570c4e8874c5181f7717.

¹ https://eel.is/c++draft/basic.compound#3.sentence-11

Pick-to: 6.9 6.8 6.5
Coverity-Id: 378348
Change-Id: Ib149386defd8b263df522a4f12b1af1b3fc1a20c
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>