moc: Add a hidden option to make QML macro warnings fatal
We don't want to make them fatal by default, as that could break user
projects with identifiers whose name collide with the macros (which is
fine if they don't actually want to use the registration macros).

However, when using qt_add_qml_module, we can assume that anything
looking like a registration macro should actually be one.

Prepare for this by adding a flag which makes the warning fatal,
and which can then be set in qtdeclarative.

Task-number: QTBUG-134148
Change-Id: I3a49968b05d66e4e4ca3a8a2dae1a7b543471c68
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>