QProgressBar: Make the default value translatable
QProgressBar currently does not expose the default value
for translation automatically. This is an issue for languages
that use a different percent formatting other than the
standard 100%. Default format has to be made explicitly
available for translation, and it is not common among the
most Qt and KDE software. This aims to fix the issue via
making it available for translation, so that it could be
translated at one source instead of fixing it in various
software.

Fixes: QTBUG-127713
Change-Id: Icd73cf48ad3cdf441e8e7aba58151cc538adff6a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>