QGIM: implement itemData/setItemData for gadgets
If the item at index is a gadget (or QObject type), then use the
properties that match the names of the roles to get and set the item
data.

This is transactional for gadgets, where we can modify a copy until all
entries have been successfully written, and then swap the copy and the
original value if all writes were successful. It cannot be transactional
for QObject instances.

Change-Id: I514853cbd67aaf6e86c100e815c8af579b8278c2
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>