CMake: Warn on creation of private modules without private headers
Private modules without private headers don't make much sense and
needlessly create CMake packages and QMake module pri files.

In qt_finalize_module we check whether the module has any private
headers. But the private module was already created before in
qt_internal_add_module. Moving the creation of the private module or the
CMake package files to the finalization phase would be a bigger
refactoring that's probably not worth the hassle.

An easier way to avoid the package creation issue is to mandate that a
Qt module must be created with NO_PRIVATE_MODULE if it doesn't have any
private headers.

Add a warning if a Qt module without private headers is created without
NO_PRIVATE_MODULE. Also warn if a Qt module with private headers is
created with NO_PRIVATE_MODULE.

This allows us to easily spot the places where NO_PRIVATE_MODULE is
missing.

Task-number: QTBUG-132526
Change-Id: I897f25790a99eaa320c39cd8c2728f26bceb824e
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>