Revert "Don't support subpixel positioning for color fonts"
This reverts commit fd9c9788f73cb088229701dd92443aa04005a4a3.

This change failed to consider the fact that color fonts may
also contain non-emoji characters and can be used for regular
text. In this case, it very much makes sense to support
subpixel positioning.

If needed, a more specific optimization can be made for emoji
characters later, but for now we just revert to fix the
regression.

Task-number: QTBUG-134626
Pick-to: 6.8 6.9
Change-Id: I07fece03d02b62811c2def26c19c5225a08a0037
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>