tst_QUuid: make big-endian runs not fail
As far as I can tell, QUuid is working correctly on big endian
platforms. But it stores the integers in such a weird fashion (I wish to
fix that for Qt 7) that this test is non-obvious.

I don't have a big-endian machine to test this with. Anyone interested
is welcome to submit patches.

Fixes: QTBUG-134634
Pick-to: 6.8 6.9
Change-Id: I2415be976d1d880d1525fffdf525dcf4f21aaa70
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>