qHashMulti*: use the seed argument in the template and noexcept traits
This shouldn't change support for existing types in Qt 6, because we
have a catch-all that calls qHash(t) and XORs the seed in. Come Qt 7,
that changes of course.

[ChangeLog][QtCore][QHash] Fixed an issue that caused qHashMulti() and
qHashMultiCommutative() to fail to compile if used with a type that
provided the required 2-argument qHash() function without providing a
default seed.

Task-number: QTBUG-126659
Fixes: QTBUG-134683
Task-number: QTBUG-134690
Change-Id: Ife742fdbd53c7bd12652fffd70f35db250db2027
Reviewed-by: Marc Mutz <marc.mutz@qt.io>