QNetworkAccessManager: protect more QElapsedTimer with isValid() checks
It might be redundant, but I can't definitely prove that these timers
are already started by the time we call elapsed() or restart() on them;
I "just" think they are because the CI is happy with the existing code,
even when adding assertions into QElapsedTimer.

As a precautionary measure, add more isValid() protections to QNAM code.
Also replace restart() with start() since the return value is unused.

Pick-to: 6.9 6.8 6.5
Change-Id: Ic64ab27116d8b1cb6c14e18dee79a15aa40844ce
Reviewed-by: Marc Mutz <marc.mutz@qt.io>