QSqlQueryModel: restore RVO in data()
The return of a default-constructed QVariant 'v' from three branches
invites NRVO, but most compilers won't let it kick in, because there
are two return statements that don't return 'v'.

Return an rvalue from each branch, so RVO is guaranteed (since C++17)
to kick in.

Amends the start of the public history.

Picking back all the way, since it's risk-free.

Pick-to: 6.9 6.8 6.5 5.15
Change-Id: I495885bee3ebba63c9e52640903c792011c1dee2
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>