QGIM: support move-only row types
A move-only type can be used as the row if the container of the rows
can be populated via insert from another container. This is the case for
the standard containers, but not for e.g. QList.

If an insert(pos, start, end) overload is available, then use that also
in the moveRows() implementation to move the data over.

The only limitation now is that we cannot assign a complete row type as
data of an item, as we cannot get a move-only value out of a const
QVariant.

Change-Id: Icea66ed82c601fd819f7db2eca2f8e023d8aa7c2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>