Cope as best we can if fractional-part and group separators coincide
The user might misconfigure their system this way. CLDR's mn_Mong_MN
even had this from v43 (when I reported it) to v45 (after which it was
fixed), so some operating system using CLDR might have that problem if
configured for this locale.

This naturally makes number parsing ambiguous in general, when the
last group might be a fractional part or the end of the whole-number
part, but we can at least catch the cases where the pattern of lengths
of groups of digits would violate grouping rules for the locale under
one reading or the other. When it is ambiguous, interpret the last
group as a fractional part. That shall sometimes be wrong, but the
system is misconfigured in this case anyway, so we're allowed to not
care.

Add some test-cases for this scenario, making clear which are
unambiguous (malformed in one reading, well-formed in the other) and
which are ambiguous (so resolved by treating last separator as the
fractional-part one).

Fixes: QTBUG-134913
Change-Id: Ib8df9bcd9ca88ad4e949565b1ecfba3ef64b587d
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>